[sitecore-jss-nextjs] reorganize exports into graphql module
