--- packages/sitecore-jss-nextjs/graphql-client.d.ts
@@ -1,1 +1,0 @@
-export * from './types/graphql-client/index';
