--- packages/sitecore-jss-nextjs/graphql-client.js
@@ -1,1 +1,0 @@
-module.exports = require('./dist/cjs/graphql-client/index');
