+++ packages/sitecore-jss-nextjs/graphql.d.ts
@@ -1,0 +1,1 @@
+export * from './types/graphql/index';
