+++ packages/sitecore-jss-nextjs/graphql.js
@@ -1,0 +1,1 @@
+module.exports = require('./dist/cjs/graphql/index');
