+++ packages/sitecore-jss-nextjs/src/utils/index.ts
@@ -5,4 +5,3 @@ export {
   resetEditorChromes,
   resolveUrl,
 } from '@sitecore-jss/sitecore-jss/utils';
-export { getEdgeProxyContentUrl } from '@sitecore-jss/sitecore-jss/graphql';
